+++ initat/tools/daemonize.py
@@ -21,7 +21,7 @@
 
 """ daemonizes a given server """
 
-from __future__ import unicode_literals, print_function
+
 
 import argparse
 import grp
@@ -39,7 +39,7 @@ import daemon
 
 def get_gid_from_name(group):
     try:
-        if type(group) in [int, long]:
+        if type(group) in [int, int]:
             gid_stuff = grp.getgrgid(group)
         else:
             gid_stuff = grp.getgrnam(group)
@@ -53,7 +53,7 @@ def get_gid_from_name(group):
 
 def get_uid_from_name(user):
     try:
-        if type(user) in [int, long]:
+        if type(user) in [int, int]:
             uid_stuff = pwd.getpwuid(user)
         else:
             uid_stuff = pwd.getpwnam(user)
